+++ src/app/home-page/home-logged-out/home.component.scss
