+++ src/app/home-page/home-logged-out/home.component.ts
@@ -18,9 +18,9 @@ import { MatDialog, MatDialogRef } from '@angular/material';
 import { HomePageService } from 'app/home-page/home-page.service';
 import { TabDirective } from 'ngx-bootstrap/tabs';
 import { Observable, Subject } from 'rxjs';
-import { User } from '../shared/swagger/model/user';
-import { TwitterService } from '../shared/twitter.service';
-import { UserQuery } from '../shared/user/user.query';
+import { User } from '../../shared/swagger/model/user';
+import { TwitterService } from '../../shared/twitter.service';
+import { UserQuery } from '../../shared/user/user.query';
 
 /**
  * Simple youtube iframe component, too simple to have its own file
