+++ src/app/home-page/logged-in-banner/logged-in-banner.component.scss
